djangoapp/requirements.txt
# Django Core
Django==5.2.8
asgiref==3.10.0
sqlparse==0.5.3

djangorestframework>=3.14,<3.15
django-cors-headers>=4.3,<4.4

# Database
psycopg2-binary==2.9.10

# Environment Variables
python-dotenv==1.2.1

drf-spectacular==0.29.0

gunicorn==23.0.0

stripe>=14.0.0,<14.2.0

djangorestframework-simplejwt>=5.5.1,<5.6

# Additional dependencies (add as needed)
# pillow==10.4.0  # For image handling
# djangorestframework==3.15.2  # For REST APIs
# celery==5.4.0  # For async tasks
